+++ plugins/brightness/BrightnessEffect.cpp
@@ -128,6 +128,6 @@ bool BrightnessEffect::enqueueGpu(cl::CommandQueue &queue, cl::Buffer &buf, cl::
     m_kernelLinear.setArg(2, h);
     m_kernelLinear.setArg(3, brightnessFactor);
     m_kernelLinear.setArg(4, contrastFactor);
-    queue.enqueueNDRangeKernel(m_kernelLinear, cl::NullRange, cl::NDRange(w, h), cl::NullRange);
+    queue.enqueueNDRangeKernel(m_kernelLinear, cl::NullRange, cl::NDRange(static_cast<size_t>(w), static_cast<size_t>(h)), cl::NullRange);
     return true;
 }
