+++ plugins/vignette/VignetteEffect.cpp
@@ -277,6 +277,6 @@ bool VignetteEffect::enqueueGpu(cl::CommandQueue &queue, cl::Buffer &buf, cl::Bu
     m_kernelLinear.setArg(12, cosA);
     m_kernelLinear.setArg(13, sinA);
 
-    queue.enqueueNDRangeKernel(m_kernelLinear, cl::NullRange, cl::NDRange(w, h), cl::NullRange);
+    queue.enqueueNDRangeKernel(m_kernelLinear, cl::NullRange, cl::NDRange(static_cast<size_t>(w), static_cast<size_t>(h)), cl::NullRange);
     return true;
 }
