+++ src/core/IInteractiveEffect.h
@@ -21,28 +21,28 @@ struct ViewportTransform {
     // Pixels of screen per pixel of source.
     float displayScale() const {
         if (imageSize.isEmpty() || viewportSize.isEmpty()) return 1.0f;
-        const float fit = std::min(static_cast<float>(viewportSize.width()) / imageSize.width(),
-                                   static_cast<float>(viewportSize.height()) / imageSize.height());
+        const float fit = std::min(static_cast<float>(viewportSize.width()) / static_cast<float>(imageSize.width()),
+                                   static_cast<float>(viewportSize.height()) / static_cast<float>(imageSize.height()));
         return fit * zoom;
     }
 
     QPointF sourceToScreen(QPointF srcPx) const {
         if (imageSize.isEmpty() || viewportSize.isEmpty()) return {};
         const float ds      = displayScale();
-        const float regionW = viewportSize.width() / ds;
-        const float regionH = viewportSize.height() / ds;
-        const float x0      = static_cast<float>(center.x()) * imageSize.width() - regionW * 0.5f;
-        const float y0      = static_cast<float>(center.y()) * imageSize.height() - regionH * 0.5f;
+        const float regionW = static_cast<float>(viewportSize.width()) / ds;
+        const float regionH = static_cast<float>(viewportSize.height()) / ds;
+        const float x0      = static_cast<float>(center.x()) * static_cast<float>(imageSize.width()) - regionW * 0.5f;
+        const float y0      = static_cast<float>(center.y()) * static_cast<float>(imageSize.height()) - regionH * 0.5f;
         return {(srcPx.x() - x0) * ds, (srcPx.y() - y0) * ds};
     }
 
     QPointF screenToSource(QPointF screenPx) const {
         if (imageSize.isEmpty() || viewportSize.isEmpty()) return {};
         const float ds      = displayScale();
-        const float regionW = viewportSize.width() / ds;
-        const float regionH = viewportSize.height() / ds;
-        const float x0      = static_cast<float>(center.x()) * imageSize.width() - regionW * 0.5f;
-        const float y0      = static_cast<float>(center.y()) * imageSize.height() - regionH * 0.5f;
+        const float regionW = static_cast<float>(viewportSize.width()) / ds;
+        const float regionH = static_cast<float>(viewportSize.height()) / ds;
+        const float x0      = static_cast<float>(center.x()) * static_cast<float>(imageSize.width()) - regionW * 0.5f;
+        const float y0      = static_cast<float>(center.y()) * static_cast<float>(imageSize.height()) - regionH * 0.5f;
         return {x0 + screenPx.x() / ds, y0 + screenPx.y() / ds};
     }
 };
